% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deciles.R
\name{deciles}
\alias{deciles}
\title{Deciles Function.}
\usage{
deciles(x)
}
\arguments{
\item{x}{is a numeric vector containing the data needed to estimate
the deciles.}
}
\value{
It returns a vector containing the values at deciles 1 to 9 for
the data.
}
\description{
This function Estimate the deciles for the data in vector x
using the Harrell-Davis estimate of the qth quantile.
}

