% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hd.R
\name{hd}
\alias{hd}
\title{hd Function}
\usage{
hd(x, q = 0.5, na.rm = TRUE)
}
\arguments{
\item{x}{is a vector containing the data to be used in the computation.}

\item{q}{is an float indicating the quantile to be computed.
the default argument is .5.}

\item{na.rm}{is a boolean value indicating whether any NA values in the data
are to be removed.}
}
\value{
It returns a single numeric value indicating the value in the data
at the quantile specified.
}
\description{
This function computes the Harrell-Davis estimate of the qth quantile.
}

