% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_func.R
\name{sample_func}
\alias{sample_func}
\title{Sample Function}
\usage{
sample_func(x, nboot = 200)
}
\arguments{
\item{x}{is a numeric vector indicating the data to be sampled.}

\item{nboot}{is the number of bootstrap samples to take.}
}
\value{
It returns an n by n matrix containing the bootstrapped samples.
}
\description{
This function allows for sampling with replacement a vector of data.
}

