% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_function.R
\name{shifthd}
\alias{shifthd}
\title{Shift Function}
\usage{
shifthd(x, y, nboot = 200, plotit = TRUE, plotop = FALSE, SEED = TRUE)
}
\arguments{
\item{x}{is a numeric vector of data, the deciles of which will be computed.}

\item{y}{is a numeric vector of data, the deciles of which will be computed.}

\item{nboot}{is a integer value specifying the number of bootstrap samples
to compute.}
}
\description{
This function computes the confidence intervals for the difference
between the deciles of two independent groups. The simultaneous 
probability coverage is .95. The Harrell-Davis estimate of
the qth quantile is used.
}

