% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shifthd.R
\name{shifthd}
\alias{shifthd}
\title{Shift Function}
\usage{
shifthd(x, y, nboot = 200, plotit = TRUE, plotop = FALSE, SEED = TRUE)
}
\arguments{
\item{x}{is a numeric vector of data, the deciles of which will be computed.}

\item{y}{is a numeric vector of data, the deciles of which will be computed.}

\item{nboot}{is a integer value specifying the number of bootstrap samples
to compute.}

\item{plotit}{is a boolean value indicating whether the shift function is
to be plotted.}

\item{plotop}{is a boolean indicating whether the actual values at the
deciles of x are to be plotted. If not, the number of the decile is plotted.}

\item{SEED}{is a boolean indicating whether a seed is to be used for
reproducible results.}
}
\value{
It rerturns a 9 by 3 matrix containing the lower and upper 
confidence intervals for the shift function, and the shift function 
estimation itself. If plotit is TRUE it also returns a plot of the 
shift function.
}
\description{
This function computes the confidence intervals for the difference
between the deciles of two independent groups. The simultaneous 
probability coverage is .95. The Harrell-Davis estimate of
the qth quantile is used.
}

